requirements.txt
numpy==1.22.0
pandas==1.2.4
tqdm==4.61.2
matplotlib==3.3.4
ipywidgets
scikit-learn
scipy
torch
torchaudio
torchinfo
tensorflow
tensorflow-io
librosa
torchinfo
rarfile
torchaudio
gdown
puts==0.0.8
SpeechRecognition
langdetect
transformers
azure-cognitiveservices-speech